+++ src/redux/store.ts
@@ -2,9 +2,9 @@
 console.log("[DEBUG] Initializing Redux store");
 import { configureStore } from '@reduxjs/toolkit';
 import portfolioReducer from './portfolioSlice';
-import marketDataReducer from '@redux/marketDataSlice';
-import tradingReducer from '@redux/tradingSlice';
-import learningReducer from '@redux/learningSlice';
+import marketDataReducer from './marketDataSlice';
+import tradingReducer from './tradingSlice';
+import learningReducer from './learningSlice';
 
 export const store = configureStore({
   reducer: {
@@ -12,8 +12,12 @@ export const store = configureStore({
     marketData: marketDataReducer,
     trading: tradingReducer,
     learning: learningReducer
-  }
+  },
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware({
+      serializableCheck: false,
+      immutableCheck: false
+    })
 });
 
 export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;